Split upload controller into a single-item upload and a csv upload controller
